+++ packages/typespec-rust/src/codegen/models.ts
@@ -595,7 +595,7 @@ function addSerDeHelper(field: rust.ModelField, serdeParams: Set<string>, format
       // if the scalar is a float, replace the . as it's illegal in an identifier
       literalValueName = literalValueName.replace('.', 'point');
     } else if (literal.valueKind.kind === 'String') {
-      literalValueName = literalValueName.replace('/', '_');
+      literalValueName = literalValueName.replace(/\W/g, '_');
     }
 
     const optional = field.optional ? 'optional_' : '';
@@ -856,7 +856,7 @@ function buildLiteralSerialize(indent: helpers.indentation, name: string, field:
     }, {
       pattern: 'false',
       body: (indent) => `${indent.get()}serializer.serialize_none()\n`,
-    }])}`
+    }])}\n`;
   } else {
     content += `${indent.get()}${toSerialize}`;
   }